requirements.txt
streamlit>=1.35.0
mysql-connector-python>=8.3.0
python-dotenv>=1.0.1
pandas>=2.2.0
